feat(lotus-sim): split info command file
